Added a lexer class.
The lexer takes a pair of iterators and a pointer to an associative
array consisting of regex, function pairs. The lexer has a get()
method which reads input until a match is found amongst the regexes.
At that point, the associated function is called with the input as
its parameter.