All unit tests are now test_suite objects.
Error reporting must be fixed. It should be kept as a data
structure in the test_suite object. Then the top level test can be
run in two steps. The first step runs the tests and compiles all
errors. The second step reports them in a format that makes sense.
This should get the std::ostream& objects out of all the functions.